updating header mismatch

+++ src/M5UnitWeightI2C.h
@@ -60,11 +60,11 @@ class M5UnitWeightI2C {
     uint8_t getEmaFilter(void);
 
     // returns an average reading; times = how many times to read
-    long read_average(byte times = 10);
+    float read_average(byte times = 10);
 
     // returns (read_average() - OFFSET), that is the current value without the
     // tare weight; times = how many readings to do
-    double get_value(byte times = 1);
+    float get_value(byte times = 1);
 
     // returns get_value() divided by SCALE, that is the raw value divided by a
     // value obtained via calibration times = how many readings to do
